+++ R/provider-claude.R
@@ -8,6 +8,8 @@ NULL
 #' Chat with an Anthropic Claude model
 #'
 #' @description
+#' `r support_badge("official")`
+#'
 #' [Anthropic](https://www.anthropic.com) provides a number of chat based models
 #' under the [Claude](https://claude.com/product/overview) moniker. Note that a
 #' Claude Pro membership does not give you the ability to call models via the
@@ -257,7 +259,6 @@ method(chat_body, ProviderAnthropic) <- function(
       )
       tools[[tool_def@name]] <- tool_def
       tool_choice <- list(type = "tool", name = tool_def@name)
-      stream <- FALSE
       output_config <- NULL
     }
   } else {
@@ -282,7 +283,15 @@ method(chat_body, ProviderAnthropic) <- function(
   tools <- as_json(provider, unname(tools))
 
   params <- chat_params(provider, provider@params)
-  if (has_name(params, "budget_tokens")) {
+
+  if (has_name(params, "reasoning_effort")) {
+    thinking <- list(type = "adaptive")
+    output_config <- modify_list(
+      output_config,
+      list(effort = params$reasoning_effort)
+    )
+    params$reasoning_effort <- NULL
+  } else if (has_name(params, "budget_tokens")) {
     thinking <- list(
       type = "enabled",
       budget_tokens = params$budget_tokens
@@ -315,7 +324,8 @@ method(chat_params, ProviderAnthropic) <- function(provider, params) {
       top_k = "top_k",
       max_tokens = "max_tokens",
       stop_sequences = "stop_sequences",
-      budget_tokens = "reasoning_tokens"
+      budget_tokens = "reasoning_tokens",
+      reasoning_effort = "reasoning_effort"
     )
   )
 
@@ -892,6 +902,14 @@ models_claude <- function(
     cache = "none"
   )
 
+  models_list(provider)
+}
+
+#' @export
+#' @rdname chat_anthropic
+models_anthropic <- models_claude
+
+method(models_list, ProviderAnthropic) <- function(provider) {
   req <- base_request(provider)
   req <- req_url_path_append(req, "/models")
   resp <- req_perform(req)
@@ -910,10 +928,6 @@ models_claude <- function(
   df <- cbind(df, match_prices("Anthropic", df$id))
   df[order(-xtfrm(df$created_at)), ]
 }
-
-#' @export
-#' @rdname chat_anthropic
-models_anthropic <- models_claude
 
 # Helpers ----------------------------------------------------------------