+++ tests/testthat/test-provider-claude.R
@@ -12,8 +12,8 @@ test_anthropic_provider <- function() {
   )
 }
 
-test_that("can make simple batch request", {
-  vcr::local_cassette("anthropic-batch")
+test_that("can make simple request", {
+  vcr::local_cassette("anthropic-basic")
 
   chat <- chat_anthropic_test("Be as terse as possible; no punctuation")
   resp <- chat$chat("What is 1 + 1?")